pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "channels>=4.0",
    "daphne>=4.2.0",
    "django>=4.2",
    "djangorestframework>=3.16.0",
    "pillow>=11.2.1",
]
